requirements.txt
absl-py==0.9.0
args==0.1.0
astor==0.8.1
attrs==19.3.0
backports.csv==1.0.7
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.12.13
botocore==1.15.13
bs4==0.0.1
bx-python==0.8.8
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clint==0.5.1
contextlib2==0.5.5
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
docutils==0.15.2
EasyProcess==0.2.10
gast==0.2.2
gensim==3.8.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.8.0
idna==2.8
importlib-metadata==1.5.0
internetarchive==1.8.5
jmespath==0.9.5
joblib==0.14.1
jsonable==0.3.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
kdap==0.1.22
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.1.3
mwapi==0.5.1
mwcli==0.0.3
mwparserfromhell==0.5.4
mwtypes==0.3.2
mwviews==0.1.0
mwxml==0.3.3
networkx==2.4
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
para==0.0.8
patool==1.12
prettytable==0.7.2
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==19.8.18
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pyunpack==0.1.2
PyYAML==5.3
regex==2020.2.20
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
schema==0.7.1
scikit-learn==0.22.2
scipy==1.4.1
sentence-transformers==0.2.5
sentencepiece==0.1.85
six==1.14.0
smart-open==1.9.0
soupsieve==2.0
tb-nightly==1.14.0a20190301
tensorboard==2.1.0
tensorflow==2.0.0a0
tensorflow-estimator==2.1.0
tensorflow-hub==0.5.0
tensorflowjs==1.2.6
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
torch==1.4.0
total-ordering==0.1.0
tqdm==4.43.0
transformers==2.3.0
urllib3==1.24.3
Werkzeug==1.0.0
wikipedia==1.4.0
wrapt==1.12.0
xmltodict==0.12.0
zipp==3.1.0
absl-py==0.9.0
args==0.1.0
astor==0.8.1
attrs==19.3.0
backports.csv==1.0.7
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.12.13
botocore==1.15.13
bs4==0.0.1
bx-python==0.8.8
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
clint==0.5.1
contextlib2==0.5.5
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
docutils==0.15.2
EasyProcess==0.2.10
gast==0.2.2
gensim==3.8.1
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.8.0
idna==2.8
importlib-metadata==1.5.0
internetarchive==1.8.5
jmespath==0.9.5
joblib==0.14.1
jsonable==0.3.1
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
kdap==0.1.22
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.2.1
matplotlib==3.1.3
mwapi==0.5.1
mwcli==0.0.3
mwparserfromhell==0.5.4
mwtypes==0.3.2
mwviews==0.1.0
mwxml==0.3.3
networkx==2.4
nltk==3.4.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
para==0.0.8
patool==1.12
prettytable==0.7.2
protobuf==3.11.3
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==19.8.18
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pyunpack==0.1.2
PyYAML==5.3
regex==2020.2.20
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
schema==0.7.1
scikit-learn==0.22.2
scipy==1.4.1
sentence-transformers==0.2.5
sentencepiece==0.1.85
six==1.14.0
smart-open==1.9.0
soupsieve==2.0
tb-nightly==1.14.0a20190301
tensorboard==2.1.0
tensorflow==2.0.0a0
tensorflow-estimator==2.1.0
tensorflow-hub==0.5.0
tensorflowjs==1.2.6
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
torch==1.4.0
total-ordering==0.1.0
tqdm==4.43.0
transformers==2.3.0
urllib3==1.24.3
Werkzeug==1.0.0
wikipedia==1.4.0
wrapt==1.12.0
xmltodict==0.12.0
zipp==3.1.0
